Drop support for all features - Only return flippers for specified features - Simpler environment configurations
