+++ app/services/feature_toggles/bundle.rb
@@ -3,8 +3,8 @@
 module FeatureToggles
   ##
   # An object responsible for getting features from the
-  # Redis cache or the PostgreSQL db if the cached features
-  # are not stored yet in Redis
+  # Redis cache or the PostgreSQL db if the feature_toggles
+  # request is not cached
   #
   # @!attribute cookie_id
   #   @return [String]
@@ -14,15 +14,16 @@ module FeatureToggles
   #   @return [Array]
   # @!attribute current_user
   #   @return [User]
-  class Factory
+  class Bundle
+    KEY_PREFIX = 'flippers'
+
     attr_reader :cookie_id, :actor, :features, :current_user
 
     ##
-    # Builds a FeatureToggles::Factory instance from given options
+    # Builds a FeatureToggles::Bundle instance from given options
     #
     # @param opts [Hash]
-    #
-    # @return [FeatureToggles::Factory] an instance of this class
+    # @return [FeatureToggles::Bundle] an instance of this class
     #
     def self.build(opts = {})
       new(opts)
@@ -36,47 +37,35 @@ module FeatureToggles
     end
 
     ##
-    # Return an array of `all` va.gov features or
-    # return a `subset` of them based on the features
-    # params that was passed to the API from the UI
+    # Return an array of va.gov features based on the
+    # features params that was passed to the endpoint
     #
     # @return [Array]
     #
-    def list
-      features.blank? ? all : subset
-    end
+    def fetch
+      return nil if features.blank?
 
-    ##
-    # Return an array of `all` va.gov features
-    #
-    # @return [Array]
-    #
-    def all
-      features_hash.each_with_object([]) do |(name, values), acc|
-        type = values.fetch('actor_type')
-        status = enabled?(name, type)
-
-        acc << body(name.camelize(:lower), status)
-        acc << body(name, status)
-      end
-    end
-
-    ##
-    # Return a subset of va.gov features based
-    # on the features params that was passed
-    # to the API from the UI
-    #
-    # @return [Array]
-    #
-    def subset
       features.each_with_object([]) do |name, acc|
-        name = name.underscore
         type = features_hash.dig(name, 'actor_type')
         status = enabled?(name, type)
 
-        acc << body(name.camelize(:lower), status)
         acc << body(name, status)
       end
+    end
+
+    ##
+    # Return a hexdigest of the combined features
+    # so that our Redis key remains short and unique
+    #
+    # @return [String]
+    #
+    def redis_key
+      return if features.blank?
+
+      joined = features&.sort&.join('_')
+      suffix = Digest::SHA1.hexdigest(joined)
+
+      "#{KEY_PREFIX}/#{suffix}"
     end
 
     ##
@@ -84,6 +73,8 @@ module FeatureToggles
     # was enabled via the Flipper UI or if the feature
     # was toggled programmatically
     #
+    # @param name [String]
+    # @param type [String]
     # @return [Boolean]
     #
     def enabled?(name, type)
@@ -95,6 +86,8 @@ module FeatureToggles
     # the basic data structure of the feature
     # returned to the caller
     #
+    # @param name [String]
+    # @param status [Boolean]
     # @return [Hash]
     #
     def body(name, status)
@@ -107,7 +100,8 @@ module FeatureToggles
     # the currently logged in user if the feature's actor
     # is set to `user`
     #
-    # @return [Flipper::Actor, User]
+    # @param type [String]
+    # @return [Flipper::Actor, User, nil]
     #
     def actor_type(type)
       case type
