+++ spec/request/v1/feature_toggles_request_spec.rb
@@ -12,8 +12,8 @@ RSpec.describe 'FeatureToggles', type: :request do
   end
 
   before do
-    allow_any_instance_of(FeatureToggles::Factory).to receive(:features_hash).and_return(all_hash)
-    allow_any_instance_of(FeatureToggles::Factory).to receive(:enabled?).and_return(true)
+    allow_any_instance_of(FeatureToggles::Bundle).to receive(:features_hash).and_return(all_hash)
+    allow_any_instance_of(FeatureToggles::Bundle).to receive(:enabled?).and_return(true)
   end
 
   describe 'GET `index`' do
@@ -23,9 +23,7 @@ RSpec.describe 'FeatureToggles', type: :request do
           'data' => {
             'type' => 'feature_toggles',
             'features' => [
-              { 'name' => 'fooBarFeature', 'value' => true },
               { 'name' => 'foo_bar_feature', 'value' => true },
-              { 'name' => 'myNewFeature', 'value' => true },
               { 'name' => 'my_new_feature', 'value' => true }
             ]
           }
@@ -33,33 +31,9 @@ RSpec.describe 'FeatureToggles', type: :request do
       end
 
       it 'returns the selected features' do
-        get '/v1/feature_toggles?features=foo_bar_feature,myNewFeature'
+        get '/v1/feature_toggles?features=foo_bar_feature,my_new_feature'
 
         expect(JSON.parse(response.body)).to eq(select_features)
-      end
-    end
-
-    context 'without feature params' do
-      let(:all_features) do
-        {
-          'data' => {
-            'type' => 'feature_toggles',
-            'features' => [
-              { 'name' => 'fooBarFeature', 'value' => true },
-              { 'name' => 'foo_bar_feature', 'value' => true },
-              { 'name' => 'myNewFeature', 'value' => true },
-              { 'name' => 'my_new_feature', 'value' => true },
-              { 'name' => 'anotherGreatFeature', 'value' => true },
-              { 'name' => 'another_great_feature', 'value' => true }
-            ]
-          }
-        }
-      end
-
-      it 'returns all features' do
-        get '/v1/feature_toggles'
-
-        expect(JSON.parse(response.body)).to eq(all_features)
       end
     end
   end
